Remove use of custom json/xml tags in generated structures
Custom tags are no longer needed due to use of custom JSON marshalers.

The tags created a limitation where response structures could not be reliably
serialized to JSON for use outside of the SDK since certain fields were either
hidden or renamed in unusable ways to restrictions on the wire protocol. For
example, fields that appeared in headers of JSON responses were marked as
omitted from the structure and could therefore not be serialized to a JSON
string if a user wanted. Removing these tags allows all data structures to be
serialized in supported formats using Go's native marshalers.

References #130