s3manager: Add support for managed uploads to S3.
This new package supports managed uploads to S3 through the
s3manager.Upload() function, which intelligently switches to a
multipart upload when it exceeds a certain part size. In this case,
the upload will be performed concurrently across a configurable
number of goroutines. The Upload() operation also allows for
non-seekable Readers to be passed in, effectively allowing
streaming uploads of unbounded payloads to S3.

Sample usage:

	func main() {
		fd, err := os.Open("path/to/file")
		if err != nil {
			panic(err)
		}

		// Create and S3 object and upload the file.
		svc := s3.New(nil)
		resp, err := s3manager.Upload(svc, &s3manager.UploadInput{
			Bucket: aws.String("bucket"),
			Key:    aws.String("path/to/file"),
			Body:   fd,
		}, nil)
		if err != nil {
			panic(err)
		}

		fmt.Println(awsutil.StringValue(resp))
	}

References #142, #207