s3manager: Fix memory leak
Before this fix, s3manager was holding onto the entire contents of the
io.Reader that it's uploading.

Each work chunk is

  type chunk struct {
    buf []byte
    num int64
  }

and the list of completed parts is []*s3.CompletedPart which include

  PartNumber *int64

When the new completed part is created using &c.num for the PartNumber,
the pointer means that the whole surrounding chunk, including the
buffer, is still live as far as the collector is concerned.

The fix is simple: copy num.