s3manager: Fix deadlock
The dispatch method of checking for errors once per loop can deadlock,
because the workers can error and exit in between when the dispatch
goroutine checks for an error and tries to enqueue the next chunk.

Instead, use a signal chan that the dispatcher and the workers can both
use to signal all the others to exit, and receive from it while doing
all channel ops.