aws/awserr: Refactored returned SDK errors, added awserr.Error errors interface
Refactored the SDK errors away from the Go error type to awserr.Error interface type. All SDK methods
will now return an awserr.Error instead of aws.APIError as a Go error. This interface is compatible
with Go error, and can be used as one. The awserr.Error interface can also be casted to other superset
error interfaces which provide access to additional error specific information.

For example request errors can be cast to awserr.RequestFailure interface to extract the StatusCode
and RequestID from the error object.

    if err != nil {
        if reqerr, ok := err.(awserr.RequestFailure); ok {
            fmt.Println("Request Error", reqerr.Code(), reqerr.Message(), reqerr.RequestID())
        }
    }

Also added a OrigErr accessor to the awserr.Error interface which returns a Go error. If not nil, this will
contain the error which caused the awserr.Error to be created, and returned. Generally this will contain
errors returned by the Go standard library, but may also be another awserr.Error.

Migrating to this update:
* All errors returned and Request.Error are awserr.Error interfaces now, Go error is no longer returned.
* No longer need to cast from error to aws.APIError. All error types are now awserr.Error interface values.
* Can Remove all references to if awserr, ok := err.(*aws.APIError); ok { ... }
* Replace all awserr.Code and awserr.Message with err.Code(), and err.Message(). Where awserr is an error cast to *aws.APIError.
* RequestID and StatusCode fields were removed from aws.APIError, and are methods provided by awserr.RequestFailure interface