s3manager: Slight modification to Upload API [Breaking Change]
In order to more consistently support the upcoming download manager,
move the s3manager.Upload() function into an s3manager.Uploader
structure. The new API is:

	u := s3manager.NewUploader(opts) // opts can be nil
	resp, err := u.Upload(&s3manager.UploadInput{...})

This changes from:

	s := s3.New(nil)
	resp, err := s3manager.Upload(s, &s3manager.UploadInput{...}, nil)

In the new API, the client is (optionally) passed through the
UploadOptions.S3 member. If S3 is nil, a default client will be used.

Note that the DefaultConcurrency and DefaultPartSize constants are
also renamed in order to be properly namespaced.