Rewrite EachPage implementation to avoid reflect [Breaking Change]
Based on a suggestion from @nightlyone we were able to remove the
use of reflection. This change subtly changes the API on the low-level
Request.EachPage() operation. This does not change the wrapped pager
functions for each operation on the service, specifically the
"*Pages()" pagination operations. The new EachPage API is:

	EachPage(fn func(data interface{}, isLastPage bool) (shouldContinue bool)) error

Note that "data" is no longer the response type structure but a simple
interface{} type. In order to use the structure, the value must be
manually asserted to the response type:

	// Get the tables in a page of data
	tables := *data.(*dynamodb.ListTablesOutput).Tables

It is also no longer possible to omit the "shouldContinue" return
value to the function.

References #247