aws: Add WriteAtBuffer for in memory usage of s3manager.Downloader
Adds aws.WriteAtBuffer to be a simple wrapper around a []byte buffer
so content can be downloaded in memory using s3manager.Downloader.

The prefered method of using s3manager.Downloader is with os.File,
but this provides a same in memory method to avoid reading from
disk after the content was written.