service/elasticbeanstalk: Add Environment Health
AWS Elastic Beanstalk enhanced health reporting monitors the status of resources in your environment and requests that your application serves to provide a more accurate measurement of environment health and performance. Health information generated by enhanced health reporting can be viewed in the AWS Management Console or with the EB command line interface, and can be published to Amazon CloudWatch.