internal/protocol/restjson: Fix struct body payloads
Operations which would put a nested field of the input struct as the body payload
were not correctly writing the payload. The struct which contained the payload and its
field in the parent struct was being included in the body. Where as only the contents
of the payload struct should of been included.

Fix #367