Add minimum length validation constraints to parameter validation
During validation the SDK will now validate the length of fields
which the service have specified, if set, must be of a mimimum length.
This isn't a breaking change because the service would of enforced the
constraints, now the SDK will provide earlier notification, and
documentation in API operation fields which require a minimum length.