aws/service: Fix default retrier to be concurrency safe
The rand.New() rand.Source returned is not concurrency safe. Instead of coping the logic from math/rand/rand.go switched to the default rand global which implements the locks.  Since this random is just used for jittler we don't need to be too critical about the randomality between runs of the jitter.

Fix #399