Refactor internal package to private
The internal package was renamed to private. `private` is a collection of packages used internally by the SDK, and is subject to have breaking changes at any time. This package is not `internal` so that if need to use this functionality, and understand breaking changes will be made, you are able to.