Fix status waiter with status code > 4xx
The waiters were ignoring any acceptor other than "error" if there was
any request error when polling for resource status. This caused waiters
like S3 WaitForObject to return immediately with an error because 404 not
found was returned. The waiters will now handle status acceptors with
errors.

Also fixed a bug with pathAll being too greedy and matching on empty
list when it should not.

Fix #497