Fix data race in EC2Metadata client initialization
The EC2Metadata client was taking a copy of http.Client's Transporter
which is not safe. This also complicated the implementation preventing
custom configurations by users. This alternative implementation is
concurrency safe and is much more flexible with custom transports and
dialers.

The EC2Metadata Timeout override can also be disabled by setting the
EC2MetadataDisableTimeoutOverride `aws.Config` flag. This flag needs to
be passed into the Session or Config prior to creating the credentials
change and EC2RoleProvider.

Fix #509