aws/types: add ability to configure the growth rate of WriteAtBuffer
Add a property, GrowthCoeff, which can be set on a WriteAtBuffer to
control the growth rate of the internal buffer. This is useful to reduce
re-allocations in cases where the input length is unknown.

NewWriteAtBuffer allows callers to pass in an internal buffer that is
pre-allocated to an expected size, skipping the need to adjust the
growth rate.