service/sts: Adds support to the AWS Security Token Service
AWS Security Token Service

This release adds support for the new STS API GetCallerIdentity, which
returns details about the credentials used to make the API call. The
details include name and account, as well as the type of entity making
the call, such as an IAM user vs. federated user.