aws/session: Enables support for Shared Config by default
Updates the SDK so by default the Session returned will be created with the
configuration loaded from the shared config file (~/.aws/config) will also be
loaded, in addition to the shared credentials file (~/.aws/config). Options
set in both the shared config, and shared credentials will be taken from the
shared credentials file. This functionality can be disabled by setting the
AWS_SDK_CONFIG_OPT_OUT environment variable.

This changes the behavior added in v1.3.0 to load the shared config by
default instead of requiring an opt in flag. This also brings the SDK in
like with the AWS SDK for Ruby to use an opt out flag with default support for
the shared config.

Related: #472, aws/aws-sdk-ruby#1257