service/s3/s3manager: Add retry download object part body (#843)
Adds the ability to automatically retried S3 object parts that fail
after the initial request response is provided. This includes scenarios
such as the network connection is terminated, or broken.

The body parts will be retried based on the S3 service clients Max
Retries configuration. For example setting the aws.Config.MaxRetries to
3 will allow the S3 Downloader to retry part body failures up to 3
times.

This feature is enabled automatically. No additional work is needed to
take advantage of it.

Fix #466