aws/ec2metadata: Add support for EC2 User Data (#872)
Adds the `GetUserData` method to `EC2Metadata` client. This allows you
to get instance user data from the `/latest/user-data` EC2 metadata
endpoint.

If the instance does not have any data set the error code
`NotFoundError` will be returned. If any data is set that value as a
string will be returned by the method.

Fix #852