aws/signer/v4: Remove logic determining if request needs to be resigned (#876)
Improves the relyability of the request signature by resigning the
request each retry. In addition logic was added to the Send request
handler chain to ensure the delay between Sign and Send will not prevent
expired signature. This also brings the SDK in line with the other AWS
SDKs in their behavior of resigning the request each retry.

Related to #822