aws/signer/v4: Add support for URL.EscapedPath to signer (#885)
Adds support for the URL.EscapedPath method added in Go1.5. This allows
you to hint to the signer and Go HTTP client what the escaped form of the
request's URI path will be. This is needed when using the AWS v4 Signer
outside of the context of the SDK on http.Requests you manage.

Also adds documentation to the signer that pre-escaping of the URI path
is needed, and suggestions how how to do this.

aws/signer/v4 TestStandaloneSign test function is an example
using the request signer outside of the SDK.

Fix #866