aws/endpoints: Add endpoint metadata to SDK (#961)
Adds Region and Endpoint metadata to the SDK. This allows you to enumerate regions and endpoint metadata based on a defined model embedded in the SDK.

This adds constants for each partition, region, and service for all Regions and Endpoints metadata. The metadata are embedded in the SDK, and is accessible as an Endpoint Resolver via `DefaultResolver`. To enumerate over the partition data case the response of EndpointResolver to a `EnumPartitions` and calls Partitions.