service/route53: Fix URL path cleaning for Route53 API requests
In SDK release v1.6.3 fix for REST protocol aboslute URL and HTTP2 broke
Route53's URL cleaning logic that removed duplicate "hostedzone"
elements from the URL. This change fixes the URL cleaning to ensure that
the modified URL RawPath is propagated to the URL's base path. RawPath
was being ignored by URL's EscapedPath method because it was not a valid
escaping of the original uncleaned Path.

Fix #1005