aws/endpoints: Add option to resolve unknown service in a partition
Adds an option to the endpoint resolver that allows resolving endpoints
for a service that is not included in the endpoints metadata. In order
for this to work with the partitions list the region must be valid. When
using a specific partition the endpoint will be resolved regardless if
the region is also valid.

This option is ignored if StrictMatching option is enabled.