aws/credentials,aws/session: Add support for AssumeRoles with MFA
Adds support for assuming IAM roles with MFA enabled. A TokenProvider
func was added to stscreds.AssumeRoleProvider that will be called each
time the role's credentials need to be refreshed. A basic
token provider that sources the MFA token from stdin as
stscreds.StdinTokenProvider.

In addition when creating a session a new session option was added,
AssumeRoleTokenProvider. The value of this field will be passed to the
stscreds.AssumeRoleProvider if the shared config enables assume role
with MFA, and the SDK will be assuming the role. This allows you to
configure the SDK via the shared config to assume a role with MFA
tokens.

Note to use the SDK with the shared config and assume role the
SharedConfigState session option must be SharedConfigEnable, or the
AWS_SDK_LOAD_CONFIG environment variable set.

Assume role with MFA is enabled via the shared config when the
serial_number field is set.

Creating an AssumeRoleProvider with MFA

    // Initial credentials loaded from SDK's default credential chain. Such as
    // the environment, shared credentials (~/.aws/credentials), or EC2 Instance
    // Role. These credentials will be used to to make the STS Assume Role API.
    sess := session.Must(session.NewSession())

    // Create the credentials from AssumeRoleProvider to assume the role
    // referenced by the "myRoleARN" ARN. Prompting for MFA token from stdin.
    creds := stscreds.NewCredentials(sess, "myRoleArn", func(p *stscreds.AssumeRoleProvider) {
	p.SerialNumber = aws.String("myTokenSerialNumber")
	p.TokenProvider = stscreds.StdinTokenProvider
    })

    // Create service client value configured for credentials
    // from assumed role.
    svc := s3.New(sess, &aws.Config{Credentials: creds})

Creating a Session with shared config enabled to assume a role with MFA

    sess := session.Must(session.NewSessionWithOptions(session.Options{
	AssumeRoleTokenProvider: stscreds.StdinTokenProvider,
	SharedConfigState:       session.SharedConfigEnable,
    }))

    // Create service client value configured for credentials
    // from assumed role.
    svc := s3.New(sess)

Fix # 842