service/s3: Fix S3 HostID to be available in S3 request error message.
Adds a new type s3.RequestFailure which exposes the S3 HostID value from
a S3 API operation response. This is helpful when you have an error with
S3, and need to contact support. Both RequestID and HostID are needed.

When logging the error message both will be logged. The HostID value can
also be retrieved by casting the returned error to a s3.RequestFailure
type.

	result, err := svc.GetObject(&s3.GetObjectInput{
		Bucket: aws.String("myBucket"),
		Key:    aws.String("myKey"),
	})
	if err != nil {
		if s3Err, ok := err.(s3.RequestFailure); ok {
			fmt.Printf("Request failed: id: %q, host: %q\n", s3Err.RequestID(), s3Err.HostID())
			return
		}
	}

Fix #1123