private/model/api: Ensure nil passed into request for paginators
Updates the Paginator and Waiter's method code generation to ensure that
if the input parameter is a nil value that the underlying __Request
method will receive the nil value. Defering any default value logic to
the __Request method. This removes the assumption at the paginator and
waiter level that the default value if input is nil is a zero value input type.

Related #1156