aws/client: Fix clients polluting handler list
Fixes the clients potentially polluting the passed in handler list with
the client's customizations. This change ensures every client always works
with a clean copy of the request handlers and it cannot polute the handlers
back upstream.

This issue was exposed in ec2metadata client created by the default
credentials chain. The client would inject request handlers, which would
polute the backing handler list array. This caused service clients that
were created from a session to not have their Send request handler in
their list. This issue was not present in service clients such as S3
which were created with a Session's ClientCoinfig because the
ClientConfig would of always created a copy of the session.

Fix #1184