aws/endpoints: Add utilities improving endpoints lookup
Adds several utilities to the endpoints packages to make looking up
partitions, regions, and services easier.

* DefaultPartitions - returns a list of Partitions. Removing the need to cast the DefaultResolver to EnumPartitions interface.
* RegionsForService - returns the regions a service is in given the partition list, partition id and service id.
* PartitionForRegion - returns the first partition that matches a given region
* Service.Regions - returns a list of regions for a service. Excluding additional meta endpoints such as "aws-global"

This change also changes the Partition, Region, Service and Endpoint
receiver methods to be value instead of pointer. This change improves
the usage of the AwsParition functions allowing the methods calls to be
used without takign the address of the returned Partition.

Fix #994