aws/corehandlers: Do request with nil Body if was NoBody value
Updates the SDK's usage of NoBody so that HTTP requests subsitute nil
for http.NoBody. Once the request is sent the Body value will be reset
back to NoBody. This ensures that the SDK sends requests will a nil Body
when no request body is intended to be sent, and the usage of
Request's HTTPRequest.Body will not be a breaking change.

Fix #1264