aws/corehandlers: Do request with nil Body if was NoBody value (#1266)
Updates the SDK's usage of NoBody so that HTTP requests substitute nil
for http.NoBody. Once the request is sent the Body value will be reset
back to NoBody. This ensures that the SDK sends requests will a nil Body
when no request body is intended to be sent, and the usage of
Request's HTTPRequest.Body will not be a breaking change.

Also moves the aws/session's custom CA bundle test to awstesting to
make the bundle and server reusable for other SDK tests.

Fix #1264