service/s3/s3manager: Fix Downloader ignoring Range get parameter
Fixes the S3 Download Manager ignoring the GetObjectInput's Range
parameter. If this parameter is provided it will force the downloader to
fallback to a single GetObject request disabling concurrency and
automatic part size gets.

This can be improved, if there is demand, for the feature to parse the
byte range value provided and bulding range gets with it.

Fix #1296