Fixed v4 Signing for requests with Host set
Go defines that for clients Request.Host overrides the Host header
to send and only if this is empty does the Request.Write method use
the value of URL.Host.

This fixes v4 signing to abide by these rules and use Request.Host
in preference to URL.Host if set.