Optimised stripExcessSpaces performance
Optimised stripExcessSpaces by using a hand rolled trim space avoiding the overhead of strings.TrimSpace and ensuring we only do a single pass on the string for both search and copy when stripping multiple spaces.

This improves the performance on the test machine from 2378 ns/op to 1528 ns/op while keeping allocs the same.

Also:
* Added a blank and only trialing spaces test cases.