private/protocol/json/jsonutil: support fractional timestamps
Some services such as ECS return timestamps as JSON numbers with fractional
parts (typically milliseconds), like 1501917228.309.

Previously, the SDK rounded all such times to the nearest second. This fixes the
SDK to use the full provided precision.  We avoid rounding by asking the JSON
library to keep numbers as strings and convert to float ourselves (calling the
same underlying function that the parser used to call).