Fixed v4 signing for unordered multi value query parameters
Remove sorting of query string values when calculating v4 signing as this is not part of the spec:
http://docs.aws.amazon.com/general/latest/gr/sigv4-create-canonical-request.html

The spec only requires the keys, not values, to be sorted which is achieved by Query.Encode().

Also fixed TestBuildCanonicalRequest, which was incorrectly testing without this important part of the build process leading to a false success.