aws/request: Remove default port from HTTP host header (#1618)
Updates the SDK to automatically remove default ports based on the URL's scheme when setting the HTTP Host header's value. This fixes an issue in the SDK's signer that would allow nonstandard ports to be used in endpoint URLs. Specifically this allows custom ports for non-standard port elastic search endpoints.

Fix #1537