aws/session: Fix bug in session.New not supporting AWS_SDK_LOAD_CONFIG
Fixes a bug in the session.New function that was not correctly sourcing
the shared configuration files' path.

5f43689 broke the session.New's handling of the shared configfuration
files when AWS_SDK_LOAD_CONFIG was enabled, because the default file
paths were moved from the envConfig to NewSessionWithOptions. This
prevented session.New from having access to these values.

Fix #
Related to awslabs/amazon-ecr-credential-helper#84