private/protocol/rest: Update REST protocol timestamp parse format
Updates the REST protocol's timestamp format to not require GMT as the
timezone value parsed. Switching to MST in the parse format string
allows the SDK to parse other timezones.

The SDK will continue to format the timezone as GMT when serialzing out
the timestamp value for REST protocol.

Related to aws/aws-sdk-go-v2#178

Fix #1937