Update CI tests to ensure all codegen changes are accounted for in PR
Updates the CI tests to ensure that any code generation changes are
accounted for in the PR, and that there were no mistaken changes made
without also running code generation. This change should also help
ensure that code generation order is stable, and there are no ordering
issues with the SDK's codegen.