private/model/api: Update client ServiceName to be based on name of service
The generated AWS service client's `ServiceName` const was being
incorrectly generated from the service's modeled EndpointPrefix not a
unique name for the service. This causes several services such as
Neptune and RDS to have the same `ServiceName` value which is incorrect.

This change prevents new services from being named incorrectly.