aws/signer: Always sign a request with the current time.
Updates the SDK's v4 request signer to always sign requests with the
current time. For the first request attempt, the request's creation time
was used in the request's signature. In edge cases this allowed the
signature to expire before the request was sent if there was significant
delay between creating the request and sending it, (e.g. rate limiting).