private/mode/api: Fix idempotencytoken members not to require validation
Fixes the SDK's usage of API operation request members marked as
idempotency tokens to not require validation. These fields will be
autopopulated by the SDK if the user does not provide a value. The SDK
was requiring the user to provide a value or disable validation to use
these APIs.