aws/ec2metadata: Fix EC2 Metadata client panic with debug logging
Fixes a panic that could occur witihin the EC2 Metadata client when both
`AWS_EC2_METADATA_DISABLED` env var is set and log level is
LogDebugWithHTTPBody. The SDK's client response body debug functionality
would panic because the Request.HTTPResponse value was not specified.

Updates the client debug loggers to check if the response is set before
attempting to log.