aws/defaults: Add default values to SDKs default http client
Updates the SDK's usage of http.DefaultClient to make a copy, and add
ResponseHeaderTimeout, and MaxIdleConnsPerHost to the copied client's
transport.

The SDK will only use the copied HTTP client if the user does not
provide a custom HTTP client. The SDK will also only attempt to copy the
client's transport if the transport is a *http.Transport value.