service/s3/s3manager: Fix uploader to check for empty part before max parts check (#2556)
Fixes the S3 Upload manager's behavior for uploading exactly MaxUploadParts * PartSize to S3. The uploader would previously return an error after the full content was uploaded, because the assert on max upload parts was occurring before the check if there were any more parts to upload.

This change fixes that logic so that the uploader first checks if there are any more chunks to upload before checking if the max number of parts to upload have been reached.

Also clarifies the uploader's behavior of MaxUploadParts, and that it must not be used to limit the uploader's total upload size. an io.LimitReader can be used for that purpose.

Fix #2557