Refactor session credentials and add tests
Refactors the Session credentials code to be in its own file and adds
unit tests for chained assume role.

Fixes the SDK's bug not correctly assuming role with
"credentials_source" for Environment and ECS roles.

Fix #2528

For testing support, added the ability to specify the Handlers the SDK
should use at the Session Options. This allows the set of handlers to be
injected from the very beginning of the sessions credential chain.