aws/session: Fix credential loading order for env and shared config
Fixes the credential loading order for environment credentials, when the
presence of an AWS_PROFILE value is also provided. The environment
credentials have precedence over the AWS_PROFILE.

If you want the profile to have precedence over the environment
credentials, (e.g when using credential_source=Environment) specify the
Profile option value when initializing the Session.

  sess, err := session.NewSessionWithOptions(session.Options{
      Profile: "myProfile",
  })

Specifying the Options.Profile value is equivalent to the AWS CLI's
--profile flag.

Fix #2727