private/model/api: Fix API doc being generated with wrong value (#2748)
Fixes the SDK's generated API documentation for structure member being
generated with the wrong documentation value when the member was
included multiple times in the model doc-2.json file, but under
different types.

The SDK was not considering the member's type when associating documentation
with the member. Only parent struct name and member names was being
used.