Make read "connection reset" errors retryable.
Commit 2a801855816c6d4cdc9116035315bbfa8cbb9836 stopped making send
request errors automatically retryable and deferred to the default
retryer to determine whether a request should be retryable.

Then in commit c3d27102088451890a9ba5be0dc1f2e0425f0fc3 a distinction
was made between read and write "connection reset" errors with the
latter now not being retryable. There is no explanation in the commmit
or the PR why the distinction was made, but this change in SDK behavior
is causing sporadic fatal errors for us in production when connecting to
AWS services such as S3 and DynamoDB.