Fix: Pagination can issue an extra HTTP request
Many services provide methods that use the Pagination API [1]
to simplify iterating over multiple pages of results.

These APIs all accept a function returning bool that allows
the user to short-circuit the operation by returning `false`:
in theory, no more pages will be requested after the user
returns `false`.

However, if the callback returns `false` before the result set
has been exhausted, an extra HTTP request will be issued.
This adds unnecessary latency to applications and costs money,
since most services bill per request.

The relevant section from the code generator is:

```
cont := true
for p.Next() && cont {
        cont = fn(p.Page().({{ .OutputRef.GoType }}), !p.HasNextPage())
}
return p.Err()
```

go evaluates the condition from left-to-right, so `p.Next()` is
evaluated (making a HTTP call if we haven't yet reached the last page)
even if `cont` is `false`.

The fix is simply to reverse the order of this conditional.

I've included the generated changes in this pull request, I'm not sure if that's what you'd like or not.

[1] https://docs.aws.amazon.com/sdk-for-go/api/aws/request/#Pagination