Fix generated errors for some JSON APIs not including message (#3089)
Fixes the SDK's generated errors to all include the `Message` member
regardless if it was modeled on the error shape. This fixes the bug
identified in #3088 where some JSON errors were not modeled with the
Message member. This caused the errors message to be dropped, and not
retrievable.

This change also updates the SDK's generated smoke test to expect the
Message member of an error unmarshaled not to be empty.

Fixes #3088